Paired file names to widget class
