+++ lib/src/views/feed/feed_screen.dart
@@ -6,10 +6,11 @@ import 'package:feed_sx/src/data/repositories/branding/branding_repository.dart'
 import 'package:feed_sx/src/sdk/branding_sdk.dart';
 import 'package:feed_sx/src/theme.dart';
 import 'package:feed_sx/src/views/feed/components/custom_app_bar.dart';
-import 'package:feed_sx/src/views/feed/components/post_actions.dart';
-import 'package:feed_sx/src/views/feed/components/post_description.dart';
-import 'package:feed_sx/src/views/feed/components/post_header.dart';
-import 'package:feed_sx/src/views/feed/components/post_media.dart';
+import 'package:feed_sx/src/views/feed/components/post/post_actions.dart';
+import 'package:feed_sx/src/views/feed/components/post/post_description.dart';
+import 'package:feed_sx/src/views/feed/components/post/post_header.dart';
+import 'package:feed_sx/src/views/feed/components/post/post_image.dart';
+import 'package:feed_sx/src/views/feed/components/post/post_widget.dart';
 import 'package:feed_sx/src/widgets/loader.dart';
 import 'package:flutter/material.dart';
 import 'package:feed_sx/src/utils/constants/ui_constants.dart';
@@ -54,28 +55,5 @@ class FeedScreen extends StatelessWidget {
           }
           return Center(child: const Loader());
         });
-  }
-}
-
-class PostWidget extends StatelessWidget {
-  const PostWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 12),
-      child: Container(
-        color: kWhiteColor,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: const [
-            PostHeader(),
-            PostDescription(),
-            PostImage(),
-            PostActions()
-          ],
-        ),
-      ),
-    );
   }
 }
